@@ -10,18 +10,22 @@
     "analyze": "cross-env ANALYZE=true next build"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "next": "^11.1.1",
+    "@emotion/cache": "latest",
+    "@emotion/react": "latest",
+    "@emotion/server": "latest",
+    "@emotion/styled": "latest",
+    "@mui/material": "^5.2.5",
+    "next": "^12.0.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.6.4",
+    "react-hook-form": "^7.22.3",
     "react-hook-form-mui": "file:.."
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^10.2.2",
-    "@types/react": "^17.0.6",
-    "@types/react-dom": "^17.0.5",
+    "@next/bundle-analyzer": "^12.0.7",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
     "cross-env": "^7.0.3",
-    "typescript": "^4.2.4"
+    "typescript": "^4.5.4"
   }
 }
