@@ -31,12 +31,9 @@
     "build": "rimraf dist && microbundle -f modern,esm,cjs --jsx React.createElement"
   },
   "peerDependencies": {
-    "@date-io/date-fns": ">= 1.3.x <2",
-    "@date-io/dayjs": ">= 1.3.x <2",
-    "@material-ui/core": "> 4.7 <5",
-    "@material-ui/icons": "> 4.7 <5",
-    "@material-ui/pickers": "> 3.2.x <4",
-    "date-fns": ">=2.7 <3",
+    "@mui/material": "> 5.x <6",
+    "@mui/icons-material": ">= 5.x <6",
+    "@mui/lab": ">= 5.x <6",
     "react": ">=17",
     "react-hook-form": ">=7.3.5",
     "typescript": ">= 4.1"
@@ -48,35 +45,39 @@
     "trailingComma": "es5"
   },
   "devDependencies": {
+    "@emotion/cache": "latest",
+    "@emotion/react": "latest",
+    "@emotion/server": "latest",
+    "@emotion/styled": "latest",
     "@babel/core": "^7.15.0",
     "@babel/preset-typescript": "^7.15.0",
-    "@date-io/date-fns": "^1.3.13",
-    "@date-io/dayjs": "^1.3.13",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/pickers": "^3.3.10",
-    "@storybook/addon-actions": "6.3.7",
-    "@storybook/addon-docs": "6.3.7",
-    "@storybook/addon-knobs": "6.3.0",
-    "@storybook/addon-links": "6.3.7",
-    "@storybook/addons": "6.3.7",
-    "@storybook/builder-webpack5": "^6.3.7",
-    "@storybook/manager-webpack5": "^6.3.7",
-    "@storybook/react": "6.3.7",
+    "@date-io/date-fns": "^2.11.0",
+    "@date-io/dayjs": "^2.11.0",
+    "@mui/material": "^5.2.5",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/lab": "^5.0.0-alpha.61",
+    "@storybook/addon-actions": "6.4.9",
+    "@storybook/addon-docs": "6.4.9",
+    "@storybook/addon-knobs": "6.4.0",
+    "@storybook/addon-links": "6.4.9",
+    "@storybook/addons": "6.4.9",
+    "@storybook/builder-webpack5": "^6.4.9",
+    "@storybook/manager-webpack5": "^6.4.9",
+    "@storybook/react": "6.4.9",
     "@types/react": "^17.0.19",
     "@types/react-dom": "^17.0.9",
     "babel-loader": "^8.2.2",
-    "date-fns": "^2.23.0",
-    "microbundle": "^0.13.3",
+    "date-fns": "^2.27.0",
+    "microbundle": "^0.14.2",
     "mutation-observer": "^1.0.3",
     "react": "^17.0.2",
     "react-docgen-typescript-loader": "^3.7.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.14.2",
+    "react-hook-form": "^7.22.3",
     "react-is": "^17.0.2",
     "ts-loader": "^9.2.5",
     "tslib": "^2.3.1",
     "typescript": "^4.4.2",
-    "webpack": "^5.51.1"
+    "webpack": "^5.65.0"
   }
 }
