@@ -13,7 +13,8 @@ module.exports = withBundleAnalyzer({
     }
     config.resolve.alias['react'] = path.resolve(__dirname, '.', 'node_modules', 'react')
     config.resolve.alias['react-dom'] = path.resolve(__dirname, '.', 'node_modules', 'react-dom')
-    config.resolve.alias['@material-ui/core'] = path.resolve(__dirname, '.', 'node_modules', '@material-ui/core')
+    config.resolve.alias['@mui/material'] = path.resolve(__dirname, '.', 'node_modules', '@mui/material')
+    config.resolve.alias['@emotion/react'] = path.resolve(__dirname, '.', 'node_modules', '@emotion/react')
     return config
   }
 })
