@@ -1,5 +1,6 @@
+// @ts-ignore
 import { FormContainer, TextFieldElement } from 'react-hook-form-mui'
-import { Button } from '@material-ui/core'
+import { Box, Button, Stack } from '@mui/material'
 import { useState } from 'react'
 
 type FormProps = {
@@ -12,19 +13,24 @@ export default function Index() {
   }
   let defaultValues: FormProps = { hallo: '' }
   return (
-    <div>
+    <Box sx={{
+      maxWidth: 300,
+      margin: '30px auto 0 auto'
+    }}>
       <FormContainer defaultValues={defaultValues} onSuccess={onSubmit}>
-        <TextFieldElement name={'hallo'} label={'Hallo'} required /><br />
-        <TextFieldElement color={'primary'} name={'primary'} label={'Primary'} /><br />
-        <TextFieldElement color={'secondary'} name={'secondary'} label={'Secondary'} />
-        <br />
-        <Button type={'submit'} variant={'contained'} color={'primary'}>Submit</Button>
+        <Stack direction={'column'}>
+          <TextFieldElement name={'hallo'} label={'Hallo'} required /><br />
+          <TextFieldElement color={'primary'} name={'primary'} label={'Primary'} /><br />
+          <TextFieldElement color={'secondary'} name={'secondary'} label={'Secondary'} />
+          <br />
+          <Button type={'submit'} variant={'contained'} color={'primary'}>Submit</Button>
+        </Stack>
       </FormContainer>
       <div>
         Data:<br />
         {JSON.stringify(values)}
       </div>
-    </div>
+    </Box>
   )
 }
 
