@@ -1,20 +1,17 @@
-import { createMuiTheme } from '@material-ui/core/styles'
-import red from '@material-ui/core/colors/red'
+import { createTheme } from '@mui/material/styles'
+import { red } from '@mui/material/colors'
 
 // Create a theme instance.
-const theme = createMuiTheme({
+const theme = createTheme({
   palette: {
     primary: {
-      main: '#ff9922'
+      main: '#556cd6'
     },
     secondary: {
       main: '#19857b'
     },
     error: {
       main: red.A400
-    },
-    background: {
-      default: '#fff'
     }
   }
 })
