@@ -1,17 +1,10 @@
-import { MuiPickersUtilsProvider } from '@material-ui/pickers'
 import React, { FC } from 'react'
-import DateFnsUtils from '@date-io/date-fns'
-import { MuiPickersUtilsProviderProps } from '@material-ui/pickers/MuiPickersUtilsProvider'
+import AdapterDateFns from '@mui/lab/AdapterDateFns'
+import LocalizationProvider, { LocalizationProviderProps } from '@mui/lab/LocalizationProvider'
 
-class LocalizedUtils extends DateFnsUtils {
-  dateFormat = 'P'
-}
-
-export type DateFnsProviderProps = FC<
-  Omit<MuiPickersUtilsProviderProps, 'utils'> & {
+export type DateFnsProviderProps = FC<Omit<LocalizationProviderProps, 'dateAdapter'> & {
   utils?: any
-}
-  >
+}>
 
 const DateFnsProvider: DateFnsProviderProps = ({
   children,
@@ -19,9 +12,9 @@ const DateFnsProvider: DateFnsProviderProps = ({
   ...props
 }) => {
   return (
-    <MuiPickersUtilsProvider {...props} utils={utils || LocalizedUtils}>
+    <LocalizationProvider dateAdapter={AdapterDateFns} {...props}>
       {children}
-    </MuiPickersUtilsProvider>
+    </LocalizationProvider>
   )
 }
 export default DateFnsProvider
