@@ -1,22 +1,8 @@
 import React, { ChangeEvent } from 'react'
-import { makeStyles } from '@material-ui/core/styles'
-import { red } from '@material-ui/core/colors'
-import { FieldError, useController } from 'react-hook-form'
-import {
-  FormControl,
-  FormControlLabel,
-  FormHelperText,
-  FormLabel,
-  Radio,
-  RadioGroup,
-  RadioProps
-} from '@material-ui/core'
+import { red } from '@mui/material/colors'
+import { Control, FieldError, useController } from 'react-hook-form'
+import { FormControl, FormControlLabel, FormHelperText, FormLabel, Radio, RadioGroup } from '@mui/material'
 
-const useStyles = makeStyles({
-  root: {
-    color: red[400]
-  }
-})
 
 export type RadioButtonGroupProps = {
   options: any[]
@@ -32,6 +18,7 @@ export type RadioButtonGroupProps = {
   onChange?: (value: any) => void
   returnObject?: boolean
   row?: boolean
+  control?: Control<any>
 }
 
 export default function RadioButtonGroup({
@@ -46,20 +33,16 @@ export default function RadioButtonGroup({
   emptyOptionLabel,
   returnObject,
   row,
+  control,
   ...rest
 }: RadioButtonGroupProps): JSX.Element {
-  const classes = useStyles()
   const { field: { value, onChange }, fieldState: { invalid, error } } = useController({
     name,
-    rules: required ? { required: 'This field is required' } : undefined
+    rules: required ? { required: 'This field is required' } : undefined,
+    control
   })
 
   helperText = error ? (typeof parseError === 'function' ? parseError(error) : error.message) : helperText
-
-  const radioProps: RadioProps = {}
-  if (invalid) {
-    radioProps.className = classes.root
-  }
 
   const onRadioChange = (event: ChangeEvent<HTMLInputElement>) => {
     const radioValue = (event.target as HTMLInputElement).value
@@ -82,7 +65,7 @@ export default function RadioButtonGroup({
                   value={value || ''}>
         {emptyOptionLabel && (
           <FormControlLabel
-            control={<Radio {...radioProps} checked={!value} />}
+            control={<Radio sx={{ color: invalid ? red.A400 : undefined }} checked={!value} />}
             label={emptyOptionLabel}
             value=""
           />
@@ -103,7 +86,7 @@ export default function RadioButtonGroup({
           )
           return (
             <FormControlLabel
-              control={<Radio {...radioProps} checked={isChecked} />}
+              control={<Radio sx={{ color: invalid ? red.A400 : undefined }} checked={isChecked} />}
               value={optionKey}
               label={option[labelKey]}
               key={optionKey}
