@@ -1,6 +1,6 @@
 import { action } from '@storybook/addon-actions'
 import { object, text } from '@storybook/addon-knobs'
-import Button from '@material-ui/core/Button'
+import { Button } from '@mui/material'
 import { CheckboxButtonGroup, CheckboxElement, FormContainer } from '../src'
 import React from 'react'
 
