@@ -1,7 +1,7 @@
 import React from 'react'
 import { FormContainer, TextFieldElement } from '../src'
 import { useForm, useWatch } from 'react-hook-form'
-import Button from '@material-ui/core/Button'
+import { Button } from '@mui/material'
 import { action } from '@storybook/addon-actions'
 
 export default {
