@@ -2,7 +2,7 @@ import * as React from 'react'
 import { FormContainer, RadioButtonGroup } from '../src'
 import { action } from '@storybook/addon-actions'
 import { object, text } from '@storybook/addon-knobs'
-import Button from '@material-ui/core/Button'
+import { Button } from '@mui/material'
 
 export default {
   title: 'RadioButtonGroup'
